requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
langchain-text-splitters
langchain-core
chromadb
openai
pypdf
rich
pandas
pydantic
python-dotenv
